+++ HISTORY.rst
@@ -1,7 +1,19 @@
 .. :changelog:
 
+
 History
 =======
+
+1.1 (2015-02-02)
+----------------
+
+* Merge ``Signal`` and ``signal`` into one class.
+* Make ``Signal`` an alias of ``signal``.
+* Make ``Signal.define`` an alias of ``signal``.
+* Fix signal support on standalone functions
+  (https://github.com/zyga/morris/issues/1)
+* Add more documentation and tests
+* Enable travis-ci.org integration
 
 1.0 (2014-09-21)
 ----------------